+++ src/abstract/constants.ts
@@ -1,0 +1,13 @@
+export abstract class Constants {
+    static readonly TAILS: string = 'TAILS';
+    static readonly HEAD: string = 'HEAD';
+    static readonly TIE: string = 'TIE';
+    static readonly YES: string = 'YES';
+    static readonly NO: string = 'NO';
+    static readonly ERROR_NO_ROLL: string = 'No rolls';
+    static readonly ERROR_NO_DICES: string = 'No dices rolled, use rollDices()';
+    static readonly CRITICAL: string = 'critical';
+    static readonly FAILURE: string = 'failure';
+    static readonly ROL: string = 'roll';
+    static readonly MODIFIED_ROL: string = 'rollModified';
+}