+++ src/entities/responseRoll.ts
@@ -1,0 +1,9 @@
+export interface ResponseRoll {
+    rolls: Array<number | string | undefined>,
+    modifiedRolls?: Array<number | string | undefined>,
+    success:string,
+    countSuccess: number,
+    countFail: number,
+    countCritical?: number,
+    countFailure?: number
+}