+++ src/index.ts
@@ -1,3 +1,1 @@
-export * from "./classes/rolling";
-export * from "./classes/rollUtils"
-export * from "./entities/dice";
+export { RpgUtils } from "./classes/rpgUtils";